functions/fetch_weather_forecast/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.3.2
